rmodtracker: make the u32 reader's high-byte shift unsigned
data_u32le assembled its value with ( byte & 0xFF ) << 24 in signed
int, undefined whenever the top byte is 0x80 or above. Present since
the reader was written; S3M's 32-bit sample fields can reach it with
crafted input, and IT truncation fuzzing (in development) trips it
immediately. Cast each term to unsigned int before shifting. The bit
pattern and every result are unchanged - value is already assigned to
an unsigned int - confirmed by byte-exact PCM over S3M, XM and MOD
spot checks. UBSan no longer reports at the site; gcc -std=c89
-pedantic -Wall -Wextra clean.

+++ libretro-common/formats/mod/rmodtracker.c
@@ -233,10 +233,10 @@ static int data_u16le( struct data *data, int offset ) {
 static unsigned int data_u32le( struct data *data, int offset ) {
 	unsigned int value = 0;
 	if( offset >= 0 && offset + 3 < data->length ) {
-		value = ( data->buffer[ offset ] & 0xFF )
-			| ( ( data->buffer[ offset + 1 ] & 0xFF ) << 8 )
-			| ( ( data->buffer[ offset + 2 ] & 0xFF ) << 16 )
-			| ( ( data->buffer[ offset + 3 ] & 0xFF ) << 24 );
+		value = ( unsigned int ) ( data->buffer[ offset ] & 0xFF )
+			| ( ( unsigned int ) ( data->buffer[ offset + 1 ] & 0xFF ) << 8 )
+			| ( ( unsigned int ) ( data->buffer[ offset + 2 ] & 0xFF ) << 16 )
+			| ( ( unsigned int ) ( data->buffer[ offset + 3 ] & 0xFF ) << 24 );
 	}
 	return value;
 }
